chore: improve frontend project structure
