+++ .env.local.examle
@@ -1,4 +1,5 @@
 COUCHDB_URL=http://couchdb:5984
+REACT_APP_API_URL=http://localhost:3000
 COUCHDB_USER=admin
 COUCHDB_PASSWORD=admin
 COUCHDB_DATABASE=car_maintenances
