+++ docker-compose.yml
@@ -43,6 +43,8 @@ services:
       context: ./frontend
       dockerfile: ../docker/frontend.Dockerfile
     container_name: frontend
+    environment:
+      - REACT_APP_API_URL=${REACT_APP_API_URL}
     ports:
       - "80:80"
     depends_on:
