+++ frontend/src/components/Login.js
@@ -1,0 +1,58 @@
+import React, { useState } from 'react';
+
+const API_URL = process.env.REACT_APP_API_URL || 'http://localhost:3000';
+
+const Login = ({ setView, setUser }) => {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+    try {
+      const response = await fetch(`${API_URL}/api/auth/login`, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        credentials: 'include',
+        body: JSON.stringify({ username, password }),
+      });
+
+      if (response.ok) {
+        await response.json();
+        alert('Login successful');
+        setUser({ username });
+        setView('home');
+      } else {
+        const errorData = await response.json();
+        alert(`Error: ${errorData.error} - ${errorData.details || ''}`);
+      }
+    } catch (error) {
+      alert('Error al iniciar sesión');
+    }
+  };
+
+  return (
+    <form onSubmit={handleLogin}>
+      <h2>Login</h2>
+      <input
+        type="text"
+        placeholder="Username"
+        value={username}
+        onChange={(e) => setUsername(e.target.value)}
+        required
+      />
+      <input
+        type="password"
+        placeholder="Password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+        required
+      />
+      <button type="submit">Login</button>
+      <button type="button" onClick={() => setView('home')}>
+        Back
+      </button>
+    </form>
+  );
+};
+
+export default Login;
